Create TesseractTest.py
A prototype for retrieving text from an image. The get_image function converts the image from BGR to RGB and changes the resolution to let Tesseract identify characters more easily. I added functions to sharpen the image that I will play around with to increase the accuracy of Tesseract as well.